requirements.txt
uvicorn
fastapi[standard]
mysql-connector-python
pandas
# typing
# pydantic
# pymongo
# dnspython
# mysqlclient
